refactor(tests): route browser hub launchers through conftest._run_app_server
Consolidate the hand-rolled uvicorn/thread lifecycle in the test_behavioral,
test_visual and test_panels_browser hub launchers onto conftest._run_app_server,
and lift the duplicated _apply_all patch-aggregator into conftest. Drops the
now-unused threading/uvicorn/_free_port/_wait_for_port imports. No behavior
change (29 browser tests green).
